+++ art/attacks/evasion/__init__.py
@@ -42,4 +42,4 @@ from art.attacks.evasion.auto_projected_gradient_descent import AutoProjectedGra
 from art.attacks.evasion.square_attack import SquareAttack
 from art.attacks.evasion.simba import SimBA
 from art.attacks.evasion.shapeshifter import ShapeShifter
-from art.attacks.evasion.imperceptible_asr.imperceptible_attack_pytorch import ImperceptibleASRPytorch
+from art.attacks.evasion.imperceptible_asr.imperceptible_asr_pytorch import ImperceptibleASRPytorch
